requirements.txt
Flask>=3.0,<4.0
line-bot-sdk>=3.0,<4.0
python-dotenv>=1.0,<2.0
